refactor: migrate cli/domain/watcher callers off object_store shim
