+++ app/watcher/vault_watcher.py
@@ -18,7 +18,7 @@ from app.services.companion_note import find_companion_by_content_hash, read_com
 from app.ingest.vault_alpha import run_vault_alpha_ingest_paths
 from app.settings.panel_actions import PanelActionMapping, load_panel_action_mappings
 from app.settings.watcher_settings import load_watcher_settings, resolve_auto_exec_enabled
-from app.store.object_store import ObjectStore
+from app.objects import ObjectStore
 from app.watcher.events import emit_watcher_run_event
 from app.write_guard import DEFAULT_WRITE_GUARD, WritesBlockedError
 from scripts.yaml_roundtrip import load_frontmatter
