+++ tests/cli/test_index_rebuild_cli.py
@@ -13,7 +13,7 @@ from app.cli import cli
 from app.components.embeddings import get_embedding_client, get_embedding_identity
 from app.db.dsn import resolve_dsn
 from app.store import object_store as legacy_store
-from app.store.object_store import DomainObject, ObjectStore
+from app.objects import DomainObject, ObjectStore
 from app.stores import get_vector_index, reset_store_backends
